requirements.txt
aiofiles==0.7.0
aioredis==1.3.1
asyncpg==0.24.0
arq==0.20
chevron==0.14.0
click==7.1.2
fastapi==0.65.2
foxglove-web==0.0.18
jinja2==2.11.3
libsass==0.21.0
misaka==2.1.1
MarkupSafe==2.0.1
phonenumbers==8.12.28
databases[postgresql]==0.5.2
pydantic[email]==1.8.2
psycopg2==2.9.1
sqlalchemy[asyncio]==1.4.23
uvicorn==0.14.0
sentry-sdk==1.3.0
tqdm==4.62.0
ua-parser==0.10.0
ujson==4.0.2
python-pdf==0.36
requests==2.26.0
python-multipart==0.0.5
